+++ build_lib.py
@@ -97,7 +97,7 @@ try:
 except Exception as e:
 
     # output build error
-    print(e)
+    print(f"Warp build error: {e}")
 
     # report error
     sys.exit(1)