+++ warp/native/builtin.h
@@ -17,7 +17,7 @@
 #if _WIN32
 #define WP_API __declspec(dllexport)
 #else
-#define WP_API
+#define WP_API __attribute__ ((visibility ("default")))
 #endif
 
 #ifdef _WIN32
@@ -82,6 +82,13 @@ struct half
     unsigned short u;
 
     CUDA_CALLABLE inline bool operator==(const half& h) const { return u == h.u; }
+
+    CUDA_CALLABLE inline half operator+=(const half& h)
+    {
+        half sum = half(float32(*this) + float32(h));
+        this->u = sum.u;
+        return *this;
+    }
 
     CUDA_CALLABLE inline operator float32() const { return float32(half_to_float(*this)); }
     CUDA_CALLABLE inline operator float64() const { return float64(half_to_float(*this)); }