+++ warp/native/warp.h
@@ -28,7 +28,7 @@ extern "C"
     WP_API void memcpy_h2d(void* context, void* dest, void* src, size_t n);
     WP_API void memcpy_d2h(void* context, void* dest, void* src, size_t n);
     WP_API void memcpy_d2d(void* context, void* dest, void* src, size_t n);
-    WP_API void memcpy_peer(void* dest_context, void* dest, void* src_context, void* src, size_t n);
+    WP_API void memcpy_peer(void* context, void* dest, void* src, size_t n);
 
     // all memsets are performed asynchronously
     WP_API void memset_host(void* dest, int value, size_t n);
@@ -82,6 +82,12 @@ extern "C"
     WP_API void array_inner_device(uint64_t a, uint64_t b, uint64_t out, int len);
     WP_API void array_sum_device(uint64_t a, uint64_t out, int len);
 
+    WP_API int cuda_driver_version();   // CUDA driver version
+    WP_API int cuda_toolkit_version();  // CUDA Toolkit version used to build Warp
+
+    WP_API int nvrtc_supported_arch_count();
+    WP_API void nvrtc_supported_archs(int* archs);
+
     WP_API int cuda_device_get_count();
     WP_API void* cuda_device_primary_context_retain(int ordinal);
     WP_API void cuda_device_primary_context_release(int ordinal);
@@ -98,6 +104,7 @@ extern "C"
     WP_API int cuda_context_get_device_ordinal(void* context);
     WP_API int cuda_context_is_primary(void* context);
     WP_API void* cuda_context_get_stream(void* context);
+    WP_API void cuda_context_set_stream(void* context, void* stream);
     WP_API int cuda_context_can_access_peer(void* context, void* peer_context);
     WP_API int cuda_context_enable_peer_access(void* context, void* peer_context);
 
@@ -107,7 +114,16 @@ extern "C"
     // return cudaError_t code
     WP_API uint64_t cuda_context_check(void* context);
 
+    WP_API void* cuda_stream_create(void* context);
+    WP_API void cuda_stream_destroy(void* context, void* stream);
+    WP_API void cuda_stream_synchronize(void* context, void* stream);
     WP_API void* cuda_stream_get_current();
+    WP_API void cuda_stream_wait_event(void* context, void* stream, void* event);
+    WP_API void cuda_stream_wait_stream(void* context, void* stream, void* other_stream, void* event);
+
+    WP_API void* cuda_event_create(void* context, unsigned flags);
+    WP_API void cuda_event_destroy(void* context, void* event);
+    WP_API void cuda_event_record(void* context, void* event, void* stream);
 
     WP_API void cuda_graph_begin_capture(void* context);
     WP_API void* cuda_graph_end_capture(void* context);