+++ warp/tests/test_all.py
@@ -42,6 +42,8 @@ import warp.tests.test_lerp
 import warp.tests.test_smoothstep
 import warp.tests.test_model
 import warp.tests.test_fast_math
+import warp.tests.test_streams
+import warp.tests.test_torch
 
 
 def register_tests(parent):
@@ -81,6 +83,8 @@ def register_tests(parent):
     tests.append(warp.tests.test_smoothstep.register(parent))
     tests.append(warp.tests.test_model.register(parent))
     tests.append(warp.tests.test_fast_math.register(parent))
+    tests.append(warp.tests.test_streams.register(parent))
+    tests.append(warp.tests.test_torch.register(parent))
 
     return tests
 
@@ -106,6 +110,7 @@ def run():
 
 
 if __name__ == '__main__':
+
     ret = run()
 
     import sys